RIA-7911 Updating hearing requirements to adjustments - notifications (#989)
* RIA-7911 Updating hearing requirements to adjustments - notifications

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* Update suppressions.xml

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

* RIA-7911 Updating functional tests

---------
