+++ config/owasp/suppressions.xml
@@ -4,7 +4,7 @@
         <notes><![CDATA[
             This vulnerability is about potential Remote Code Execution when serializing and deserializing Java classes
             using HttpInvokerServiceExport and org.springframework.remoting.
-            As we don't use those constructs, we are not affected by it.
+            As we do not use those constructs, we are not affected by it.
             The suppression will be a long-term one. An expiry to the suppression is kept to allow re-evaluating whether
             we're still unaffected by it.
         ]]></notes>
